move PrimaryGenerator base class into Generators library
