+++ SimCore/src/SimCore/Simulator.cxx
@@ -23,7 +23,7 @@
 #include "SimCore/G4Session.h"
 #include "SimCore/G4User/TrackingAction.h"
 #include "SimCore/Geo/ParserFactory.h"
-#include "SimCore/PrimaryGenerator.h"
+#include "SimCore/Generators/PrimaryGenerator.h"
 #include "SimCore/SDs/SensitiveDetector.h"
 #include "SimCore/UserEventInformation.h"
 #include "SimCore/XsecBiasingOperator.h"
